+++ src/DeviceInterface/AndroidInterface.js
@@ -1,5 +1,5 @@
 import InterfacePrototype from './InterfacePrototype.js'
-import {autofillEnabled, sendAndWaitForAnswer, notifyWebApp} from '../autofill-utils.js'
+import {autofillEnabled, sendAndWaitForAnswer} from '../autofill-utils.js'
 import { NativeUIController } from '../UI/controllers/NativeUIController.js'
 import {processConfig} from '@duckduckgo/content-scope-scripts/src/apple-utils'
 
@@ -45,15 +45,6 @@ class AndroidInterface extends InterfacePrototype {
 
     }
 
-    postInit () {
-        const cleanup = this.scanner.init()
-        this.addLogoutListener(() => {
-            cleanup()
-            if (this.globalConfig.isDDGDomain) {
-                notifyWebApp({ deviceSignedIn: {value: false} })
-            }
-        })
-    }
     /**
      * Used by the email web app
      * Settings page displays data of the logged in user data
