+++ swift-package/Resources/assets/autofill.js
@@ -3688,21 +3688,6 @@ class AndroidInterface extends _InterfacePrototype.default {
   }
 
   async setupAutofill() {}
-
-  postInit() {
-    const cleanup = this.scanner.init();
-    this.addLogoutListener(() => {
-      cleanup();
-
-      if (this.globalConfig.isDDGDomain) {
-        (0, _autofillUtils.notifyWebApp)({
-          deviceSignedIn: {
-            value: false
-          }
-        });
-      }
-    });
-  }
   /**
    * Used by the email web app
    * Settings page displays data of the logged in user data
@@ -3893,25 +3878,6 @@ class AppleDeviceInterface extends _InterfacePrototype.default {
         await this.getAddresses();
       }
     }
-  }
-
-  async postInit() {
-    if (this.isDeviceSignedIn()) {
-      this.scanner.forms.forEach(form => form.redecorateAllInputs());
-    }
-
-    const cleanup = this.scanner.init();
-    this.addLogoutListener(() => {
-      cleanup();
-
-      if (this.globalConfig.isDDGDomain) {
-        (0, _autofillUtils.notifyWebApp)({
-          deviceSignedIn: {
-            value: false
-          }
-        });
-      }
-    });
   }
   /**
    * Used by the email web app
@@ -4993,8 +4959,6 @@ class InterfacePrototype {
     await this.settings.refresh();
   }
 
-  postInit() {}
-
   async isEnabled() {
     return (0, _autofillUtils.autofillEnabled)(this.globalConfig);
   }
@@ -5017,6 +4981,21 @@ class InterfacePrototype {
       window.addEventListener('load', handler);
       document.addEventListener('readystatechange', handler);
     }
+  }
+
+  postInit() {
+    const cleanup = this.scanner.init();
+    this.addLogoutListener(() => {
+      cleanup();
+
+      if (this.globalConfig.isDDGDomain) {
+        (0, _autofillUtils.notifyWebApp)({
+          deviceSignedIn: {
+            value: false
+          }
+        });
+      }
+    });
   }
   /**
    * @deprecated This was a port from the macOS implementation so the API may not be suitable for all
@@ -5604,8 +5583,7 @@ class WindowsInterface extends _InterfacePrototype.default {
   }
 
   postInit() {
-    const cleanup = this.scanner.init();
-    this.addLogoutListener(cleanup);
+    super.postInit();
     this.ready = true;
   }
 
@@ -13515,7 +13493,7 @@ class ExtensionTransport extends _index.DeviceApiTransport {
     }
 
     if (deviceApiCall instanceof _deviceApiCalls.SetIncontextSignupDismissedAtCall) {
-      return deviceApiCall.result(await extensionSpecificSetIncontextSignupDismissedAt(deviceApiCall.params.value));
+      return deviceApiCall.result(await extensionSpecificSetIncontextSignupDismissedAt(deviceApiCall.params));
     }
 
     if (deviceApiCall instanceof _deviceApiCalls.GetIncontextSignupDismissedAtCall) {
@@ -13524,7 +13502,7 @@ class ExtensionTransport extends _index.DeviceApiTransport {
 
 
     if (deviceApiCall instanceof _deviceApiCalls.SendJSPixelCall) {
-      return deviceApiCall.result(await extensionSpecificSendPixel(deviceApiCall.params.pixelName));
+      return deviceApiCall.result(await extensionSpecificSendPixel(deviceApiCall.params));
     }
 
     throw new Error('not implemented yet for ' + deviceApiCall.method);
@@ -13597,17 +13575,15 @@ async function getContentScopeConfig() {
   });
 }
 /**
- * @param {import('../__generated__/validators-ts').SendJSPixelParams['pixelName']} pixelName
+ * @param {import('../__generated__/validators-ts').SendJSPixelParams} params
  */
 
 
-async function extensionSpecificSendPixel(pixelName) {
+async function extensionSpecificSendPixel(params) {
   return new Promise(resolve => {
     chrome.runtime.sendMessage({
       messageType: 'sendJSPixel',
-      options: {
-        pixelName
-      }
+      options: params
     }, () => {
       resolve(true);
     });
@@ -13623,14 +13599,16 @@ async function extensionSpecificGetIncontextSignupDismissedAt() {
     });
   });
 }
+/**
+ * @param {import('../__generated__/validators-ts').SetIncontextSignupDismissedAt} params
+ */
 
-async function extensionSpecificSetIncontextSignupDismissedAt(value) {
+
+async function extensionSpecificSetIncontextSignupDismissedAt(params) {
   return new Promise(resolve => {
     chrome.runtime.sendMessage({
       messageType: 'setIncontextSignupDismissedAt',
-      options: {
-        value
-      }
+      options: params
     }, () => {
       resolve(true);
     });
